RungeKutta.jl

This is the documentation of RungeKutta.jl, a Julia package implementing Runge-Kutta methods.

Manual

RungeKutta.jl

Manual

API

Types

Functions

Solvers

Utilities

Index

API

All exported types and functions are considered part of the public API and thus documented in this manual.

Types

RungeKutta.ButcherTableau — Type

ButcherTableau

A composite type for the Butcher tableau of a RungeKuttaSolver.

Constructors

ButcherTableau(A, b, c, s, p[, d, q])
ButcherTableau(tableau)

Arguments

A :: AbstractMatrix : matrix of coefficients.

b :: AbstractVector : vector of weights.

c :: AbstractVector : vector of nodes.

s :: Integer : number of stages.

p :: Integer : order of accuracy.

d :: AbstractVector : embedding's vector of weights.

q :: Integer : embedding's order of accuracy.

tableau :: AbstractMatrix : matrix from which all other fields are automatically constructed, with structure:

\[\begin{array}{c|c} c & A \\ \hline p & b \\ q & d \end{array}\]

Functions

show : shows name and contents.

summary : shows name.

RungeKutta.RungeKuttaSolver — Type

An abstract type for Runge-Kutta solvers.

RungeKutta.ExplicitRungeKuttaSolver — Type

ExplicitRungeKuttaSolver <: RungeKuttaSolver

A composite type for explicit RungeKuttaSolvers.

Constructors

ExplicitRungeKuttaSolver(tableau, h[, adaptive])
ERK(args...; kwargs...)

Arguments

tableau :: ButcherTableau : Butcher tableau.

stepsize :: StepSize : step-size.

adaptive :: AdaptiveParameters : embedded method's parameters.

Functions

show : shows name and contents.

summary : shows name.

RungeKutta.ImplicitRungeKuttaSolver — Type

ImplicitRungeKuttaSolver <: RungeKuttaSolver

A composite type for implicit RungeKuttaSolvers.

Constructors

ImplicitRungeKuttaSolver(tableau, h, newton[, adaptive])
IRK(args...; kwargs...)

Arguments

tableau :: ButcherTableau : Butcher tableau.

stepsize :: StepSize : step-size.

newton :: NewtonParameters : simplified Newton's parameters.

adaptive :: AdaptiveParameters : embedded method's parameters.

Functions

show : shows name and contents.

summary : shows name.

RungeKutta.RungeKuttaSolution — Type

RungeKuttaSolution <: InitialValueSolution

A composite type for the solution of an InitialValueProblem obtained with a RungeKuttaSolver.

Constructors

RungeKuttaSolution(u, t[, k])
RungeKuttaSolution(problem, solver; save_stages = false)

Arguments

u : numerical solution.

t : time grid.

k : stages.

problem :: InitialValueProblem.

solver :: RungeKuttaSolver.

save_stages :: Bool : flags when to save all stages into k.

Functions

getindex : get value(s) and time.

lastindex : last index.

length : number of time steps.

setindex! : set value(s) and time.

show : shows name and contents.

size : number of variables and time steps.

summary : shows name.

Functions

NSDEBase.solve — Function

solve(problem::InitialValueProblem, solver::RungeKuttaSolver; save_stages::Bool = false) :: RungeKuttaSolution

returns the RungeKuttaSolution of an InitialValueProblem. save_stages flags when to save all stages into solution.k.

NSDEBase.solve! — Function

solve!(solution::RungeKuttaSolution, problem::InitialValueProblem, solver::RungeKuttaSolver) :: RungeKuttaSolution

returns the RungeKuttaSolution of an InitialValueProblem.

Solvers

Explicit

RungeKutta.Euler — Function

Euler(; h = 0.0) :: ExplicitRungeKuttaSolver
ExplicitEuler(args...; kwargs...) :: ExplicitRungeKuttaSolver

returns an ExplicitRungeKuttaSolver for the 1st-order Euler method. ```

RungeKutta.Midpoint — Function

Midpoint(; h = 0.0) :: ExplicitRungeKuttaSolver
ExplicitMidpoint(args...; kwargs...) :: ExplicitRungeKuttaSolver

returns an ExplicitRungeKuttaSolver for the 2nd-order midpoint method.

RungeKutta.Heun2 — Function

Heun2(; h = 0.0) :: ExplicitRungeKuttaSolver

returns an ExplicitRungeKuttaSolver for the 2nd-order Heun method.

RungeKutta.Ralston2 — Function

Ralston2(; h = 0.0) :: ExplicitRungeKuttaSolver

returns an ExplicitRungeKuttaSolver for the 2nd-order Ralston method.

RungeKutta.Heun3 — Function

Heun3(; h = 0.0) :: ExplicitRungeKuttaSolver

returns an ExplicitRungeKuttaSolver for the 3rd-order Heun method.

RungeKutta.Kutta3 — Function

Kutta3(; h = 0.0) :: ExplicitRungeKuttaSolver

returns an ExplicitRungeKuttaSolver for the 3rd-order Kutta method.

RungeKutta.Ralston3 — Function

Ralston3(; h = 0.0) :: ExplicitRungeKuttaSolver

returns an ExplicitRungeKuttaSolver for the 3rd-order Ralston method.

RungeKutta.SSPRK3 — Function

SSPRK3(; h = 0.0) :: ExplicitRungeKuttaSolver

returns an ExplicitRungeKuttaSolver for the 3rd-order Strong-Stability-Preserving Runge-Kutta method.

RungeKutta.RK4 — Function

RK4(; h = 0.0) :: ExplicitRungeKuttaSolver

returns an ExplicitRungeKuttaSolver for the 4th-order Runge-Kutta method.

RungeKutta.Rule38 — Function

Rule38(; h = 0.0) :: ExplicitRungeKuttaSolver

returns an ExplicitRungeKuttaSolver for the 4th-order 3/8-rule method.

RungeKutta.HeunEuler — Function

HeunEuler(; h = 0.0, δ = 0.0, ϵ = 1e-5, K = 100) :: ExplicitRungeKuttaSolver

returns an ExplicitRungeKuttaSolver for the 2nd-order Heun-Euler method with 1st-order error estimate.

RungeKutta.Fehlberg45 — Function

Fehlberg45(; h = 0.0, δ = 0.0, ϵ = 1e-5, K = 100) :: ExplicitRungeKuttaSolver
F45(args...; kwargs...) :: ExplicitRungeKuttaSolver

returns an ExplicitRungeKuttaSolver for the 4th-order Fehlberg method with 5th-order error estimate.

RungeKutta.DormandPrince54 — Function

DormandPrince54(; h = 0.0, δ = 0.0, ϵ = 1e-5, K = 100) :: ExplicitRungeKuttaSolver
DP54(args...; kwargs...) :: ExplicitRungeKuttaSolver

returns an ExplicitRungeKuttaSolver for the 5th-order Dormand-Prince method with 4th-order error estimate.

RungeKutta.Verner65 — Function

Verner65(; h = 0.0, δ = 0.0, ϵ = 1e-5, K = 100) :: ExplicitRungeKuttaSolver
V65(args...; kwargs...) :: ExplicitRungeKuttaSolver

returns an ExplicitRungeKuttaSolver for the 6th-order Verner method with 5th-order error estimate.

Implicit

RungeKutta.BackwardEuler — Function

BackwardEuler(; h = 0.0, ϵ = 1e-3, K = 10) :: ImplicitRungeKuttaSolver
ImplicitEuler(args...; kwargs...) :: ImplicitRungeKuttaSolver

returns an ImplicitRungeKuttaSolver for the 1st-order backward Euler method.

RungeKutta.ImplicitMidpoint — Function

ImplicitMidpoint(; h = 0.0, ϵ = 1e-3, K = 10) :: ImplicitRungeKuttaSolver

returns an ImplicitRungeKuttaSolver for the 2nd-order implicit midpoint method.

RungeKutta.CrankNicolson — Function

CrankNicolson(; h = 0.0, ϵ = 1e-3, K = 10) :: ImplicitRungeKuttaSolver

returns an ImplicitRungeKuttaSolver for the 2nd-order Crank-Nicolson method.

RungeKutta.SDIRK3 — Function

SDIRK3(; h = 0.0, ϵ = 1e-3, K = 10) :: ImplicitRungeKuttaSolver

returns an ImplicitRungeKuttaSolver for the 3rd-order SDIRK method.

RungeKutta.GaussLegendre4 — Function

GaussLegendre4(; h = 0.0, ϵ = 1e-3, K = 10) :: ImplicitRungeKuttaSolver
GL4(args...; kwargs...) :: ImplicitRungeKuttaSolver

returns an ImplicitRungeKuttaSolver for the 4th-order Gauss-Legendre method.

RungeKutta.GaussLegendre6 — Function

GaussLegendre6(; h = 0.0, ϵ = 1e-3, K = 10) :: ImplicitRungeKuttaSolver
GL6(args...; kwargs...) :: ImplicitRungeKuttaSolver

returns an ImplicitRungeKuttaSolver for the 6th-order Gauss–Legendre method.

RungeKutta.LobattoIIIA4 — Function

LobattoIIIA4(; h = 0.0, ϵ = 1e-3, K = 10) :: ImplicitRungeKuttaSolver

returns an ImplicitRungeKuttaSolver for the 4th-order Lobatto IIIA method.

RungeKutta.LobattoIIIB2 — Function

LobattoIIIB2(; h = 0.0, ϵ = 1e-3, K = 10) :: ImplicitRungeKuttaSolver

returns an ImplicitRungeKuttaSolver for the 2nd-order Lobatto IIIB method.

RungeKutta.LobattoIIIB4 — Function

LobattoIIIB4(; h = 0.0, ϵ = 1e-3, K = 10) :: ImplicitRungeKuttaSolver

returns an ImplicitRungeKuttaSolver for the 4th-order Lobatto IIIB method.

RungeKutta.LobattoIIIC2 — Function

LobattoIIIC2(; h = 0.0, ϵ = 1e-3, K = 10) :: ImplicitRungeKuttaSolver

returns an ImplicitRungeKuttaSolver for the 2nd-order Lobatto IIIC method.

RungeKutta.LobattoIIIC4 — Function

LobattoIIIC4(; h = 0.0, ϵ = 1e-3, K = 10) :: ImplicitRungeKuttaSolver

returns an ImplicitRungeKuttaSolver for the 4th-order Lobatto IIIC method.

RungeKutta.RadauIA3 — Function

RadauIA3(; h = 0.0, ϵ = 1e-3, K = 10) :: ImplicitRungeKuttaSolver

returns an ImplicitRungeKuttaSolver for the 3rd-order Radau IA method.

RungeKutta.RadauIA5 — Function

RadauIA5(; h = 0.0, ϵ = 1e-3, K = 10) :: ImplicitRungeKuttaSolver

returns an ImplicitRungeKuttaSolver for the 5th-order Radau IA method.

RungeKutta.RadauIIA3 — Function

RadauIIA3(; h = 0.0, ϵ = 1e-3, K = 10) :: ImplicitRungeKuttaSolver

returns an ImplicitRungeKuttaSolver for the 3rd-order Radau IIA method.

RungeKutta.RadauIIA5 — Function

RadauIIA5(; h = 0.0, ϵ = 1e-3, K = 10) :: ImplicitRungeKuttaSolver

returns an ImplicitRungeKuttaSolver for the 5th-order Radau IIA method.

Utilities

RungeKutta.ℛ — Function

ℛ(z::Number, solver::RungeKuttaSolver) :: Number

returns the stability function of a RungeKuttaSolver:

\[ R(z) = \frac{\det(I - z(A - \mathbb{1}b^\intercal))}{\det(I - zA)}.\]

ℛ(Z::AbstractMatrix, tableau::ButcherTableau) :: AbstractMatrix

returns the stability function of a RungeKuttaSolver.

Index

RungeKutta.ButcherTableau

RungeKutta.ExplicitRungeKuttaSolver

RungeKutta.ImplicitRungeKuttaSolver

RungeKutta.RungeKuttaSolution

RungeKutta.RungeKuttaSolver

NSDEBase.solve

NSDEBase.solve!

RungeKutta.BackwardEuler

RungeKutta.CrankNicolson

RungeKutta.DormandPrince54

RungeKutta.Euler

RungeKutta.Fehlberg45

RungeKutta.GaussLegendre4

RungeKutta.GaussLegendre6

RungeKutta.Heun2

RungeKutta.Heun3

RungeKutta.HeunEuler

RungeKutta.ImplicitMidpoint

RungeKutta.Kutta3

RungeKutta.LobattoIIIA4

RungeKutta.LobattoIIIB2

RungeKutta.LobattoIIIB4

RungeKutta.LobattoIIIC2

RungeKutta.LobattoIIIC4

RungeKutta.Midpoint

RungeKutta.RK4

RungeKutta.RadauIA3

RungeKutta.RadauIA5

RungeKutta.RadauIIA3

RungeKutta.RadauIIA5

RungeKutta.Ralston2

RungeKutta.Ralston3

RungeKutta.Rule38

RungeKutta.SDIRK3

RungeKutta.SSPRK3

RungeKutta.Verner65

RungeKutta.ℛ